requirements.txt
langchain>=0.1.0
langgraph>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.28.0
qdrant-client>=1.6.0
pypdf>=3.17.0
langsmith>=0.0.40